worker/requirements.txt
redis==7.4.0
python-dotenv

